refactor: types updated
